refactor: Split useSnakeGame.ts — separate constants, canvas renderer, and input handlers
